Changed Mesh to inherit from VertexArray, removed 'dynamic' from examples
