+++ examples/dotstim_demo.py
@@ -9,8 +9,7 @@ theta = random(n_points) * np.pi * 2
 verts = np.vstack((np.sin(theta) * width, (random(n_points) - .5) * height, np.cos(theta) * width)).T
 
 cylinder = rc.Mesh.from_incomplete_data(verts, drawmode=rc.Mesh.points, position=(0, 0, -2),
-                                        point_size=1, dynamic=True,
-                                        mean_center=False)
+                                        point_size=1, mean_center=False)
 cylinder.uniforms['diffuse'] = 1., 1., 1.
 cylinder.uniforms['flat_shading'] = True
