+++ ratcave/mesh.py
@@ -8,12 +8,10 @@ from .utils import vertices as vertutils
 from .utils import NameLabelMixin, BindingContextMixin, BindNoTargetMixin, BindTargetMixin, create_opengl_object, vec
 from . import physical, shader
 from .texture import Texture
-from .vertex import VertexBuffer, ElementArrayBuffer
+from .vertex import VertexArray
 import pyglet.gl as gl
 from copy import deepcopy
 from sys import platform
-
-from warnings import warn
 
 
 def gen_fullscreen_quad(name='FullScreenQuad'):
@@ -36,14 +34,14 @@ class EmptyEntity(shader.HasUniformsUpdater, physical.PhysicalGraph, NameLabelMi
         pass
 
 
-class Mesh(shader.HasUniformsUpdater, physical.PhysicalGraph, NameLabelMixin, BindingContextMixin, BindNoTargetMixin):
+class Mesh(shader.HasUniformsUpdater, physical.PhysicalGraph, NameLabelMixin, BindingContextMixin, BindNoTargetMixin, VertexArray):
     triangles = gl.GL_TRIANGLES
     points = gl.GL_POINTS
 
     bindfun = gl.glBindVertexArray if platform != 'darwin' else gl.glBindVertexArrayAPPLE
 
     def __init__(self, arrays, textures=(), mean_center=True,
-                 gl_states=(), drawmode=gl.GL_TRIANGLES, point_size=15, visible=True, **kwargs):
+                 gl_states=(), point_size=15, visible=True, **kwargs):
         """
         Returns a Mesh object, containing the position, rotation, and color info of an OpenGL Mesh.
 
@@ -64,20 +62,19 @@ class Mesh(shader.HasUniformsUpdater, physical.PhysicalGraph, NameLabelMixin, Bi
         Returns:
             Mesh instance
         """
-
-        super(Mesh, self).__init__(**kwargs)
+        # Change vertices from an Nx3 to an Nx4 array by appending ones.  This makes some calculations more efficient.
+        arrays = list(arrays)
+        arrays[0] = np.append(arrays[0], np.ones((arrays[0].shape[0], 1), dtype=np.float32), axis=1)
+        # arrays, indices = vertutils.reindex_vertices(self.arrays)  # Indexing
+        super(Mesh, self).__init__(arrays=arrays, **kwargs)
         self.reset_uniforms()
 
-        arrays = tuple(np.array(array, dtype=np.float32) for array in arrays)
-        # self.arrays, self.array_indices = arrays, None  # No indexing
-        self.arrays, self.indices = vertutils.reindex_vertices(arrays)  # Indexing
-
         # Mean-center vertices and move position to vertex mean.
-        # vertex_mean = self.arrays[0].mean(axis=0)  # No indexing
-        vertex_mean = self.arrays[0][self.indices, :].mean(axis=0)  # Indexing
+        vertex_mean = self.arrays[0][:, :3].mean(axis=0)  # No indexing
+        # vertex_mean = self.arrays[0][self.indices, :].mean(axis=0)  # Indexing
 
         if mean_center:
-            self.arrays[0][:] -= vertex_mean
+            self.arrays[0][:, :3] -= vertex_mean
         if 'position' in kwargs:
             self.position.xyz = kwargs['position']
         elif mean_center:
@@ -85,18 +82,11 @@ class Mesh(shader.HasUniformsUpdater, physical.PhysicalGraph, NameLabelMixin, Bi
         self._mean_center = mean_center
         # self.position.xyz = vertex_mean if not 'position' in kwargs else kwargs['position']
 
-        # Change vertices from an Nx3 to an Nx4 array by appending ones.  This makes some calculations more efficient.
-        arrays = list(self.arrays)
-        arrays[0] = np.append(self.arrays[0], np.ones((self.arrays[0].shape[0], 1), dtype=np.float32), axis=1)
-        self.arrays = arrays
-
         self.textures = list(textures)
         self.gl_states = gl_states
-        self.drawmode = drawmode
         self.point_size = point_size
         self.visible = visible
         self.id = None
-        self._loaded = False
 
     def __repr__(self):
         return "<Mesh(name='{self.name}', position_rel={self.position}, position_glob={self.position_global}, rotation={self.rotation})".format(
@@ -105,7 +95,7 @@ class Mesh(shader.HasUniformsUpdater, physical.PhysicalGraph, NameLabelMixin, Bi
     def copy(self):
         """Returns a copy of the Mesh."""
         return Mesh(arrays=deepcopy([arr.copy() for arr in [self.vertices, self.normals, self.texcoords]]),
-                    texture=self.textures, mean_center=deepcopy(self._mean_center),
+                    textures=self.textures, mean_center=deepcopy(self._mean_center),
                     position=self.position.xyz, rotation=self.rotation.__class__(*self.rotation[:]),
                     scale=self.scale.xyz,
                     drawmode=self.drawmode, point_size=self.point_size, visible=self.visible,
@@ -193,25 +183,6 @@ class Mesh(shader.HasUniformsUpdater, physical.PhysicalGraph, NameLabelMixin, Bi
                                                                                                  dtype=np.float32)
         return cls(arrays=(vertices, normals, texcoords), **kwargs)
 
-    def load_vertex_array(self):
-        self.id = create_opengl_object(gl.glGenVertexArrays if platform != 'darwin' else gl.glGenVertexArraysAPPLE)
-        with self:
-            for loc, verts in enumerate(self.arrays):
-                vbo = verts.view(type=VertexBuffer)
-                with vbo:
-                    gl.glVertexAttribPointer(loc, verts.shape[1], gl.GL_FLOAT, gl.GL_FALSE, 0, 0)
-                    gl.glEnableVertexAttribArray(loc)
-                self.arrays[loc] = vbo
-        self._loaded = True
-
-    @property
-    def indices(self):
-        return self._indices if hasattr(self, '_indices') else None
-
-    @indices.setter
-    def indices(self, int_array):
-        self._indices = int_array.view(type=ElementArrayBuffer)
-
     def draw(self):
         """ Draw the Mesh if it's visible, from the perspective of the camera and lit by the light. The function sends the uniforms"""
         if not self._loaded:
@@ -225,12 +196,7 @@ class Mesh(shader.HasUniformsUpdater, physical.PhysicalGraph, NameLabelMixin, Bi
                 texture.bind()
 
             self.uniforms.send()
-            with self:
-                if self.indices is None:
-                    gl.glDrawArrays(self.drawmode, 0, self.arrays[0].shape[0])
-                else:
-                    with self.indices as indices:
-                        gl.glDrawElements(self.drawmode, indices.shape[0], gl.GL_UNSIGNED_INT, 0)
+            super(Mesh, self).draw()
 
             for texture in self.textures:
                 texture.unbind()
